+++ examples/02_typescript/src/Item.tsx
@@ -1,18 +1,18 @@
 import React, { useState } from 'react';
 
-import { createResource, useResource } from 'react-hooks-fetch';
+import { createAsync, useAsync } from 'react-hooks-fetch';
 
 import DisplayData from './DisplayData';
 
-const initialResource = createResource<{ data: { first_name: string } }>('https://reqres.in/api/users/1?delay=3');
+const initialResult = createAsync<{ data: { first_name: string } }>('https://reqres.in/api/users/1?delay=3');
 
 const Item: React.FC = () => {
   const [id, setId] = useState('1');
-  const resource = useResource(initialResource);
+  const result = useAsync(initialResult);
   return (
     <div>
       User ID: <input value={id} onChange={e => setId(e.target.value)} />
-      <DisplayData id={id} resource={resource} />
+      <DisplayData id={id} result={result} />
     </div>
   );
 };